requirements.txt
flask
python-dotenv
pymysql
flask-wtf
flask-login
openai
paramiko
openpyxl
cryptography
Werkzeug
requests
xlsxwriter
xlrd
pandas
scikit-learn
jieba
lifelines
httpx
python-pptx
